Updated imports and moved GameData classes to their respective domain package.
