+++ src/main/java/com/langleon/dsobuildsim/gamedata/GameDataConfig.java
@@ -4,11 +4,10 @@ import com.langleon.dsobuildsim.buffs.PhysicDefinition;
 import com.langleon.dsobuildsim.buffs.TonicDefinition;
 import com.langleon.dsobuildsim.character.CharacterClass;
 import com.langleon.dsobuildsim.dragonstones.DragonStoneDefinition;
-import com.langleon.dsobuildsim.dto.ClassStatsDTO;
+import com.langleon.dsobuildsim.gamedata.dto.ClassStatsDTO;
 import com.langleon.dsobuildsim.enchantments.EnchantmentDefinition;
 import com.langleon.dsobuildsim.essences.EssenceDefinition;
 import com.langleon.dsobuildsim.gems.GemDefinition;
-import com.langleon.dsobuildsim.items.core.LevelMultiplierTable;
 import com.langleon.dsobuildsim.items.mythicitems.MythicItemDefinition;
 import com.langleon.dsobuildsim.items.setitems.SetItemDefinition;
 import com.langleon.dsobuildsim.items.uniqueitems.UniqueItemDefinition;
